vals/Modularizar y mejorar validaciones
