+++ client/src/helpers/post-form-validator.js
@@ -1,20 +1,9 @@
-// VALID ATTRIBUTES :
-export const validAttr = {
-  name_on_doc: {
-    maxLength: 60,
-    minLength: 3,
-  },
-  number_on_doc: {
-    maxLength: 50,
-    minLength: 3,
-  },
-  comments: {
-    maxLength: 800,
-    minLength: 0,
-  },
-};
+import { stringContainsURLs } from "./generic-validators";
+import { isStringBetweenXAndYCharsLong } from "./generic-validators";
+import { isValidDateBoolean } from "./generic-validators";
 
-// POST FORM VALIDATOR :
+//* -------- POST FORM VALIDATOR :  ---------- --------------------------
+
 export function postFormValidator(post, t) {
   try {
     const { name_on_doc, number_on_doc, country_found, date_found, comments } =
@@ -55,61 +44,11 @@ export function postFormValidator(post, t) {
     if (!date_found) {
       throw new Error(t("postForm.errorDateFalsy"));
     }
-    if (!checkDate(date_found)) {
+    if (!isValidDateBoolean(date_found)) {
       throw new Error(t("postForm.errorDateInvalid"));
     }
     return true;
   } catch (error) {
     return { error: error.message };
-  }
-}
-
-// check Date :
-function checkDate(date_found) {
-  let dateNow = Date.now();
-  let dateFoundParsed = Date.parse(date_found);
-  if (typeof dateFoundParsed !== "number") {
-    console.log("typeof dateFoundParsed !== 'number' ");
-    return false;
-  }
-  if (dateNow < dateFoundParsed) {
-    console.log("Error: La dateFound no puede ser mayor que la fecha actual.");
-    return false;
-  }
-  return true;
-}
-
-// IS STRING BETWEEN X AND Y CHARACTERS LONG:
-export function isStringBetweenXAndYCharsLong(x, y, argumento) {
-  let error = `The argument "x" must be a positive number`;
-  if (!x || typeof x !== "number" || x < 1) {
-    throw new Error(`The argument "x" must be a positive integer number`);
-  }
-  let minCharsLong = x;
-  if (!y || typeof y !== "number" || y < 1) {
-    throw new Error(`The argument "y" must be a positive integer number.`);
-  }
-  let maxCharsLong = y;
-  if (
-    typeof argumento === "string" &&
-    argumento.length >= minCharsLong &&
-    argumento.length <= maxCharsLong
-  ) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-// STRING CONTAINS URLS :
-export function stringContainsURLs(argumento) {
-  if (
-    new RegExp(
-      "([a-zA-Z0-9]+://)?([a-zA-Z0-9_]+:[a-zA-Z0-9_]+@)?([a-zA-Z0-9.-]+\\.[A-Za-z]{2,4})(:[0-9]+)?(/.*)?"
-    ).test(argumento)
-  ) {
-    return true;
-  } else {
-    return false;
   }
 }
